Move functionality from TrajectoryFollower to URCommander
